Fix alias of promoted column (#1006)
When promoting a non string column, the tag promoter was wrongly applying the expression alias tags[asdasd] to both the external function (correct) and to the nested column (wrong). This means we had one alias shadowing the other and the resulting formatted expression was replacing the column reference with the alias, which resulted in an incorrect query.

This applies the expression alias only to the correct expression.