Add Missing time bucketed column on discover (#1018)
The discover dataset was missing the configuration for the bucketed column time.
Before the AST, since column_expr was relying on the individual datasets (events and transactions) to do the time bucketing. That was a violation of the query processing pipeline (as long as time bucketing is a dataset query processor not a storage one as it is today).
This means discover needs to provide the bucketing spec, that is then translated into a storage specific column name.