feat(migrations): Initial migration for migrations storage (#997)
This creates the basic migration, operations, and runner implementation. It also
provides a first migration for the migrations storage (a ClickHouse table).

`run_migration()` runs a single migration given its ID and inserts a row with
the ID and a completed status in the migrations table.

Only single node clusters are currently supported.