feat(migrations): Add the not started status to the migrations table
We should support the not started status in the migrations table in
case we need to rollback a migration for any reason. Since we cannot
actually delete the row from the table we should just update the status
to not started.

This could have been implemented as a second migration, however since no
one is running these yet, I think it would be cleaner to simply edit
the initial migration.