feat(migrations): Set status to in progress before running migration
Set the migration status to in progress before running the migration
operations.

Note after the version column is introduced we will need to increment
the version on every status update.

We add a hack that allows writing system migrations (except those with
completed status) to fail with unknown table error. This is required
since the table to store the migrations might not exist yet.