feat(migrations): Add not started status and version to migrations table (#1029)
We should support the not started status in the migrations table in
case we need to rollback a migration for any reason. Since we cannot
actually delete the row from the table we should just update the status
to not started.

Also adds a version column to the table which will be used to determine
the remaining row on merge instead of DateTime which only has a one
second resolution.