fix(ast) Apply aliases to literals during formatting (#1020)
Literals as well can be assigned aliases during the query processing. An easy example is for transactions:
https://github.com/getsentry/snuba/blob/master/snuba/datasets/transactions.py#L54

I think I just forgot to apply them to the formatted query during formatting. This adds them
Also, while I am here, also de-clutter aliases a bit by not applying them to columns when the name and the aliases (before escaping) coincide. This is makes the formatted query easier to read.