Count columns properly (#1038)
The ast can contain aliased expression for columns (and we do that during processing). Technically the old representation could as well, but all processing happened after the splitter in column_expr methods.

The column splitter needs to know how many columns we are loading to apply its optimization. In that case (col as A) and (col as B) are the same since clickhouse is loading one column not 2.
In the ast these are different nodes because of the alias attribute, so we need to collapse those two when counting.