feat(migrations): Move the forward method into migration (#1031)
Migrations can now customize their own forwards and backwards
methods. In the majority of cases, migrations will inherit from
MultiStepMigration and the default forwards and backwards methods
will be used. However it allows us to provide slightly different behavior
that is useful in some cases, such as the for the migrations that bootstrap
the migration system itself.