ref: Skip non-error outcomes in snuba (#1008)
For billing of attachments and performance features, we will soon start
to emit outcomes for attachments and sessions, as well as mark
transactions specially.

Since TSDB and Sentry cannot distinguish these yet, we chose to skip
outcomes and only retain the ones for classic error tracking events.
This will make sure that attachments do not show up in any stats and
interfere with spike protection. The final decision whether we should
skip transactions is still pending in the billing team, but we will
likely want to exclude them.