fix(splitter) Do not apply the column splitter if the resulting query is invalid (#1040)
The column splitter tries to split it, but the minimal query generated is invalid because it still contains a reference to the project alias in the orderby, but it looses the declaration of project that was in the select when we build the minimal select.

This adds a method to validate the aliases referenced in a query, and applies it to the column splitter. If the resulting query is invalid, the splitter gives up and go to the next one.