Rollout simple event query (#1046)
Group.filter_by_event_id and api.group-hashes are the simplest queries on the events dataset. These are not customizable, they do not contain tags, and are not being split. They are in low volume as well.
These seem the best candidates to start with to rollout the AST on the events dataset. Which has some meaningful translation and a lot more query processors than outcomes.