Make old tags representation prefix (#1041)
This is basically throw-away code. snuba/datasets/tags_column_processor.py will be removed as soon as the AST is rolled out.
This PR turns some of the expressions from the [] notation to the array access function notation since the AST expressions are formatted using array function instead of the subscriptable notation.
This is meant to help validate that the AST generated expression is actually equivalent to the legacy one by making the two easier to compare. they would be the same string now.