feat(migrations): Add basic support for replicated tables
When a merge tree table engine is specified, the replicated version
will be created instead of the non replicated one, as long as the
cluster that the migration is being performed on has single_node = False
specified.

Since it's difficult to convert a table from MergeTree to ReplicatedMergeTree
later, it's better to assume that the user wants the replicated variant
if they have specified they are operating a multi node cluster, even if
there only happens to be one replica specified in configuration.