feat(ast) Expand alias references at the end of parsing to hide them to query processing (#1024)
To make the AST model consistent we need to expand alias references at the end of parsing (so they disappear) and are reintroduced at the end before formatting.
That's what this PR does:

- it finds all aliases declared in the query
- it replaces alias references with the expanded expressions and as a bonus enforces no cycles are created (which before would have been rejected by clickhouse).

The way query processing works does not change. A query processor decides whether they want to just apply a pure transformation function to the whole query (which is by far the most common case and where they structurally cannot introduce alias shadowing) or, for more advanced processing they have to mindful of aliases.