feat(query) Getting rid of aliases step 1: Introduce SelectedExpression in the Query class (#1057)
We want to remove aliases from query processing. This is the plan:
The user provides a query that may define and reference aliases.
During parsing we keep track of the names of the fields the user expects in the request and we expand the alias references so the query is fully expanded.
The Whole query processing/translation deals with a fully expanded tree without aliases.
At the end, when we are about to format the query, we reintroduce aliases for the clickhouse query (internally chosen names that cannot conflict with anything in the query). These are used for mapping the SELECT expressions in the output with the original names provided by the user and to compact replicated subtrees.

This is the step in this PR:
If we remove aliases from the AST we still need to keep track somewhere of the names of the fields the user expects in the QueryResult structure. This is now part of the Query class and kept outside of the AST by introducing the SelectedExpression class. Functionally nothing really happens since those names are not yet used and we still rely on the aliases in the AST. See next step