feat(migrations): Make storage set string instead of enum
This is a follow up to conversation in
https://github.com/getsentry/snuba/pull/1058#discussion_r447353998.

We want to support the possibility of storage sets being deprecated in
the future, therefore we should not reference the storage set enum
directly. Instead the migration operations provide the storage set
as a string. If the string does not match a currently valid storage
set, then the migration is skipped.