ref(datasets): Remove querylog and internal only datasets
Remove the querylog dataset as well as the ability to set datasets
as "internal". Since querylog data cannot be accessed via API, it
should not have an associated dataset.

Previously a dataset acted like a storage and was necessary in order
to have a consumer and migrations for a specific storage. This isn't
the case anymore, and we no longer need to create a dataset in order
to create and fill a table.