ref(optimize): Optimize targets a storage instead of dataset
The optimize command optimizes a storage instead of a dataset. Since
there may be more than one storage associated with a dataset now, it
allows a specific storage to be targeted. All of the table schemas
associated with the storage will be optimized not just the one defined
as the "writable" schema.

The dataset option is temporarily supported, as we need to update
production scripts before removing this, however should be removed as a
follow up.