feat(migrations): Add support for marking migrations as deprecated
A migration can now be marked as deprecated. If a migration needs to
become deprecated in future, this will be done by editing the previously
created migration file to deprecated = True. If a migration is deprecated,
it will be reversed before any other migrations are run when running all.

This allows us to avoid unnecessarily running every deprecated migration
in scenarios where we are running them all in sequence from the start
(such as a new Snuba installation of bringing up a new node). It also
ensures that nodes are brought back to a consistent state if the
deprecated migration was previously run.

Once a migration has been marked as deprecated, it can also be safely
deleted in a subsequent version of Snuba, if we wish to remove it from
the codebase entirely.

About deletion of entire migration groups: The process described here
provides the ability to target individual migrations within a group
for deprecation. If we want to drop an entire migration group (as is
proposed in #1074) we do not necessaraily need to reverse out all of
the migrations of that group. We can just add a final migration that
cleans up any relevant tables then deprecate/delete the group completely.