ref(storages): Split events_ro into its own storage (#1105)
The `sentry_dist` vs `sentry_dist_ro` table selection is now
the responsibility of the dataset storage selector and does not
rely on a query processor to switch out the table name.
This is now implemented for the Discover dataset. Dataset storage
selection will also need to be rolled out to the Events dataset before
the ReadOnlyTableSelector can be removed.

Currently (in this PR) our new `event_ro` storage  is still part of the
events storage set. In a follow up, it will be moved to its own storage
set so that it can have a distinct cluster configuration. In prod, the
`sentry_dist_ro` table points at a different ClickHouse cluster to
`sentry_dist` so this is necessary to accurately reflect our current
configuration in settings.