feat(migrations): Add migrations for events table (#1103)
Add the migration that creates the events table.

It collapses all of the existing add column statements that
are part of the events migration function since those migrations
should have already been performed in a previous Snuba version.

The operation to create the sentry_dist_ro table should also be part of
this migration, however it needs to be split into a separate storage and
storage set first since in prod the distributed tables reference different
machines, i.e. a different ClickHouse cluster.