feat(migrations): Add migration for onpremise compatibility
This migration provides compatibility for any onpremise user that
is migrating from a pre- migration system version of Snuba. We need
to sync up all of the columns that have changed in the sentry_local
individually to ensure all of these get the correct state going
forward.

We would need to provide something similar for the other two tables
(transactions and errors) that have had their schema changed since
they were first introduced.