Enforce AST parsing (#1136)
I did not observe, in production, any warning from failure in parsing the AST that were not also fatal errors when running the query. Earlier on we were used to have these cases and the query could still safely run on the legacy representations but I did not observe any case like this in more than a month.