feat(migration) Allow the old DDL system to skip materialization during CREATE TABLE (#1150)
Clickhouse has some restrictions on materialized columns during table creation. For example a materialized column cannot reference a nested column (ClickHouse/ClickHouse#12586). Likely this is because the nested column is represented as multiple arrays on the schema (and thus on the materialization statement), though upon creation, those multiple arrays are still one single nested column.
Such columns have to be either skipped and added after the CREATE statement with an ALTER statement or created without materialization and updated afterwards.

The old DDL/migration system would not support this. The schema (used for querying and processing) is what generates the CREATE statement. So, if we remove the column from the schema so it is not created, a lot of stuff breaks on the query path, the invariant that the schema represented what is on the DB would not be valid. So this PR takes the opposite approach allowing the old migration system to delay the creation of specific columns (present in the schema) that have to be added with an ALTER statement.