feat(perf) Add _tags_hash_map column to the events table (#1159)
Tags hash map is an optimization meant to replace _flattened_tags.
Instead of concatenating the tags into one string and filtering through the LIKE function, this creates a hash of the pairs key=value and stores them in an array.
Filtering is efficient since we can add a bloom filter index on the array INDEX bf_tags_hash_map _tags_hash_map TYPE bloom_filter() GRANULARITY 1. This is very quick at reducing the size of the dataset when the query filters specific tags without unpacking the tags column.

This adds the column to the events dataset. It populates it through materialization (this makes it easy to manage replacements, since no change is needed). It also adds a test that ensures the column is populated properly.