feat(querylog) Move querylog logic into its own module and use an enum for the status (#1163)
This is a preliminary refactoring to make it easier to log invalid requests in a safe way.
It moves the querylog recording logic into its own module (that logic also takes care of recording the query status metrics). This makes it cleaner to add methods to record invalid queries (during parsing) which may happen outside the web module.
I don't think there is a need to keep that logic coupled to the web module, it does not depend on it and it can be agnostic.

Also I added an enum for the query status, since now we do not have a Clickhouse enum for that.