ref(storage-schemas): Remove the storage schemas class (#1170)
The StorageSchemas code was migrated over from DatasetSchemas when datasets contained multiple storages and thus had multiple schemas. A storage is now associated with a single schema. The parts that are only required for DDL are also removed from the storage class as these will no longer be needed with the new migration system.