feat(migrations): Add unsharded option to MergeTree tables
Add an unsharded option for merge tree tables that replicates the same
table across all shards of a cluster instead of actually sharding the
table. This will be needed for the grouped message and group assignee
tables.

Updates the path to use the storage set name instead of requiring the
user to provide a "layer" macro. Also updates the format of the path
to make it more readable in Zookeeper.

This is not quite the format we have in prod on our existing tables.
Since our previously defined paths cannot be changed, we will need to
add some escape hatch where we can override these with our actual path
values later.