Cleanup logger that we do not need anymore (#1175)
This metric was collected one month ago to verify how much we use alias references in Snuba query.
This code was temporary and now we expand aliased expressions, so the alias references are not in the query anymore at the point this metric is logged. Thus we do not need this anymore.