fix: Add request context indexing
Discover and issue search currently offer three fields, `http.method`,
`http.referer`, and `http.url`. All of these fields are mapped to
`contexts[http.*]` by the column resolver, however none of these fields
currently work.

In errors, the http_method and http_referer fields are indexed as
promoted context data, but are not in the promoted_context mapping and
so they are never found by context expression expansion. For
transactions, the http fields were simply not indexed.

I've added the http.url field to the events dataset, and added both the
request and http contexts to errors and transactions. The errors dataset
has the request interface indexed as `request` so I've included that key
on transactions with the goal of ending up on that alias when errors
goes live.