feat(slo) Introduce a one pass query validator applied after parsing. (#1182)
Adds the infrastructure to run multiple expression validators on a query in one single pass over the AST.
Each validator is an instance of the class ExpressionValidator and it validates one individual expression.
The context each validator has is just the expression and the dataset.
One of this validator validates function calls, which is the most common use case. Each function is validated by one specific FunctionCallValidator.
There is a sequence of function validators applied to all queries (well, there will be, now the sequence is empty) and each dataset can provides its own function validators to be applied as well.