feat(slo) Add validator for LIKE / NOT LIKE functions. (#1183)
This validates all functions like and notLike. This enforces that LIKE/NOT LIKE can only be used on string types (which is what clickhouse allows. No Fixed strings no Datetime).
This is done by Adding a reusable SignatureValidator that validates the types of the parameters provided.

Limitations: it does not address expressions like f(column) LIKE 'asdasd'. We would need proper type inference based on clickhouse type checking to allow for that. It would be glorious, but also arguably over engineered.