feat(hackweek) Remove merged jsonschema and validate it is downloaded properly. (#1202)
Removes the jsonschema initially committed since we download the schema during make develop and avoid accidentally merge the schema downloaded during make.

Also runs mypy right after downloading the schema. If the schema is not there, malformed or in the wrong place mypy fails with code 2. Then make will fail instead of silently be ok until we run it later.