fix(validator) Minimal fix for `like` validator in case of arrayjoin (#1207)
Unfortunately validating an expression in isolation against the table schema is not enough.
If an arrayjoin clause is passed on column x. Such column (that was an array(string)) becomes a string, thus like on that column must now work.

This is a quick fix for production. I will do a proper one later that takes the arrayjoin into account without being overly permissive.