feat(split): Skip column splitting if condition on ID column
Do not perform column splitting if the ID column is already present in
the where clause since the purpose of the column splitter is to
add an event ID condition and this is unlikely to yield any benefit.

This scenario comes up frequently when we are looking up a single event
by its project and event ID. Splitting should not be applied in this
case.