feat(dataset) Snuba Custom Functions (#1223)
Introduces a dataset query processor that is supposed to make it easy to define custom functions for a dataset.
The idea is that a custom function is defined by, guess what? a name, a signature with a list of parameter names and a body. The body is a string that contains an infix notation expression.
This expression is parsed into the corresponding AST expression upon instantiation.