feat(migrations): Add a command to reverse a migration
A completed or in progress migration (without a subsequent completed
migration) can be reversed by the user.

This command is not intended to be used in most normal scenarios as
some data migrations cannot be fully reversed. It is provided primarily
to support recovery of database state in some failure scenarios, such as
if we need to revert a change quickly, or to recover a migration that
has failed halfway and needs to be restarted.