fix(migrations): Fix migration to recreate transactions table
When recreating the transactions table we also need to check that the
primary key is the latest one and update it if it isn't, as this has
changed since the table was first introduced.

Not updating the primary key would lead to both an incorrect primary key
as well as a failure in the subsequent migration when the _finish_date
column (which was previously part of the primary key) would be dropped.