feat(processors) Add http fields to event/transaction/errors processor
Standardize how the http fields are parsed in the processors and add the
http method and referer to the columns.

This also removes the request field from the contexts object in the errors
processor, but all the data that was in the context is now either in a column or
in the tags.