feat(search) Add functions for creating conditions on nullable arrays
The current IN conditions we use on `exception_stacks_mechanism.handled`
are not what we want in the product. This field has an unconventional
data model where undefined/None mean the same thing as True, and not
False. This means that end user queries like `error.handled:0` and
`error.handled:1` do not work as required. These new functions make it
possible for sentry to generate the correct conditions.

I was not able to use the `CustomFunction` abstraction for this because
it cannot currently parse the required lambda inside arrayExists().

Refs https://www.notion.so/sentry/error-handled-conditions-03bea1a3ad3d45639fc3976465638ccf