feat(tags) Tags hash map processor first step (#1259)
This is the first implementation of the query processor that uses the tags hashmap column we created a few weeks ago (#1190).

It transforms tags conditions that are in the form of tags.value[indexOf(tags.key, 'my_tag')] = 'my_val' into `has(_tags_hash_map, cityHash64('my_tag=my_val'))`

It is applied only to the transactions storage. If it works I will add to the events as well.