feat(discover): Discover table selection prioritizes top level condition
If the user passed a top level condition that allows us to unambiguously
determine the table to be used, we now use that ahead of checking columns
in the query and inferring a table.

This fixes a bug that caused the events table to be used in some casees
even when type = transaction is specified, that was introduced fairly
recently when we switched over to the AST implementation of table detection.

This potentially also caused some impossible queries from not being logged
- for example a query with a condition 'type = transaction' which
referenced events columns would not be captured as impossible though it
should be.