feat(migrations): Rebuild groupedmessages table if old version (#1281)
The primary key on the groupedmessage table was changed in
https://github.com/getsentry/snuba/pull/453. Users that had a
version of this table prior to September 2019 will still have
a missing project ID column and an incorrect primary key. This
migration fixes the schema. The migration throws an error if an
incorrect primary key is found and the user has data in the table.
This shouldn't be the case since this table should never have been
filled in onpremise, single tenant and development environments.