feat(performance) Add query profile data to ClickhouseQueryMetadata (#1292)
I am trying to collect some profile data from the query AST to store in the querylog and add in discover.
This should make it easier to analyze specific query types and understand what is slow and why.
The issue is that in querylog we only have the SQL string. Pull aggregate information from there is really hard, and in discover we only have referrer and dataset.

This one adds:

-query time range in days
-list of columns in the filter
-list of nested columns in the filter
-group by columns
-array join columns.
The profile is not yet 100% accurate, but we can use this as a start.