fix(migrations): Ensure onpremise migrations are skipped if multi node
Add a check to the migrations that recreate the transactions and
groupedmessage table to ensure that they are skipped if the relevant
cluster is not a single node cluster. These migrations are only
required for onpremise users running single node clusters upgrading
from older versions of Snuba. They are not required and likely will
not work properly in multi node environments.