feat(errors): Stop writing org ID
This PR stops writing org ID to the errors table.

Until we have a guarantee that the org ID will not change, there
isn't any value in writing it as we can never rely on the value written
there.

The benefit of not writing this value is that we may be able to start
running the errors replacer again and start using this table for a subset
of Discover queries. Previously it was too slow because we were unable
to provide an org_id in the replacements query. If it is guaranteed
be 0, we can pass that into the query to address the replacer perfomance
issues.