feat(entities) Restructure to add Entity class
Restructure the Dataset code to add a new Entity class that contains most of the
logic that was contained in the Dataset. Most of this code is simply copying the
code from the Dataset class to the related Entity. There are shims around most
of the Dataset functions to forward calls to Dataset to its contained Entity.

I also used this opportunity to remove the TimeSeriesDataset. The logic there
has been moved to TimeSeriesProcessor and added to all the relevant entities.