feat(migrations): Add get_status function
Add the ability to get the status and timestamp of a migration.
This is intended to provide a way to potentially run different code
depending on whether a migration was completed yet or when it was
completed. We did something similar with a hardcoed date for the flattened
tags and contexts optimization in transactions - however in future we should
get the date the migration was actually run rather than assume that it's
completed at the same date in all environments.