ref: `snuba cleanup` supports passing a storage name
We should run cleanup on a storage rather than a dataset. This would
allow us to remove the errors dataset, which is not being used
anywhere else except for as a mechanism to allow us to cleanup the
errors storage.

Passing --dataset is still temporarily supported while we migrate
all usages of this script to pass --storage instead.