feat(measurements) Default measurements nested column to NULL instead of 0 (#1328)
Clickhouse returns the default for the type of a column when accessing an element that does not exist on an array. This is already happening for tags where empty string is returned. Since measurements.value is not Nullable, the default returned is 0. Which is a valid measurement value so this is not ok.

This provides a new translator for nested column expressions like tags or measurements that defaults to NULL instead of defaulting to the default value of the column type.
The expression looks like
`if(has(measurements.key, 'my_key'), arrayValue(measurements.values, indexOf(measurements.key, 'my_key'), NULL)`

It has some performance downside, but it seems very small after testing it on tags.