Remove legacy query from splitter 1 (#1341)
Three of the four metrics we collected around discrepancies in managing the legacy query on query splitter turned out to be expected. So removing that log and the related dependency.

The fourth seems fine (actually it was a bug in the legacy system) but I need to do one last verification, thus logging the stack trace.