feat(spans) Experimental: Adds a potential spans table to test ingestion and joins (#1299)
This adds a span table to run some experiments and study the feasibility of supporting spans search.

This is a new storage that is optional and disabled. People running migrations will not be creating this table by default.
It does not include a query path since the storage is not referenced in any dataset.