test: Refactor tests
This PR introduces BaseStorageTest to make it easier to test a storage
without an accompanying dataset (not used yet but intended to allow for
the deletion of the errors dataset). It also simplifies BaseTest and
ensures that migrations run before every test that inherits from this,
not just those for which a dataset name is passed in the setup_method.