feat(measurements) Arrayjoin optimizer for measurements (#1245)
When we use arrayJoin on tags we need to do some additional processing on those collumns:

if we are arrayJoining on both key and value, we must arrayJoin on the pairs and not do that inndependently. arrayJoin(arrayMap(key, val -> (key, val), tags.key, tags.value) otherwise we would obtain the cartesian product of the two columns.
we optimize fitlering over the arrayJoin by doing an arrayFilter before the arrayJoin to reduce the result size.
This applies such processing to measurements.