feat(errors): Simplify schema, turn off errors migrations via old system (#1353)
This stops migrating the errors table via the old migration system,
as we will be rebuilding this table soon. Keeping migrations in both
the old and the new system in sync is something I want to avoid doing
in this case. This table is currently unused.

It also removes org_id from the required_column list since this column
will be dropped in the new version of the table.