First pass at Entities (#1320)
Restructure the Dataset code to add a new Entity class that contains most of the
logic that was contained in the Dataset. Most of this code is simply copying the
code from the Dataset class to the related Entity. There are shims around most
of the Dataset functions to forward calls to Dataset to its contained Entity.

I also used this opportunity to remove the TimeSeriesDataset. There is still logic 
around time conditions that is copied between all the entities but that should be
trivial to remove.