ref: Replacements work on a single table
Not sure if there's a use case for this that I'm not aware of, but
we never currently perform replacements that read rows from one table
and replace them in another. This removes the ability to pass a
different write and read schema to a replacer processor, which may cause
problems since the replacer assumes the same columns are present in
each.