feat(ast) Remove the alias parameter from the Matching api (#1347)
We should never try to match an alias with anything (whether it is a string or not).
Aliases are an implementation detail of query processing that is meant to ensure the clickhouse query is not ambiguous. As such they may be replaced and changed in any way by the query processing pipeline while keeping the query intact.
They are not a structural part of the query we should rely on to identify constructs or patterns.

This removes the parameter entirely from the matcher api since using it would mean we are doing something probably broken. It also makes the API more concise.