fix(entities) Remove get_abstract_data_model from Dataset
This function should exist only on the Entity, so remove it from the Dataset and
pass Entities into the functions that require access to the data model. Also
move function validators to the entity. They are not used at the Dataset level
right now. Finally move EntityKey into the factory since they are used together.