ref: Move migrations specific column types to migrations module
I debated whether this made more sense grouped together with the rest of
the ClickHouse column types as they are now, or together with
migrations. Ultimately, since these are not supposed to be used anywhere
outside of migrations, I think moving them to migrations will help avoid
confusion or accidentally using these in storage schemas.