test: Deprecate dataset_manager (#1384)
`dataset_manager` existed in a time when migrations were tied
to a specific dataset, as was each test suite. It was responsible for
running the migrations for that dataset then cleaning up those tables
afterwards. We don't need this anymore, since now all migrations are run
(not just ones specific to the dataset under test) and all storages
truncated between every test.