feat(discover): Move Discover storage selection to entity selection
This moves the Discover storage selection decision up one level making
it the responsibility of the Discover dataset to perform entity selection.
This is closer to how Discover will eventually work - a specific entity
(errors, transactions, or both) will need to be specified by the user
when using SnQL. In order to use SnQL, this logic would also need to be
ported to Sentry.

This works by making the `Dataset.get_entity()` method responsible for entity
selection and providing an alternative implementation for Discover.
This seems broadly compatible with the direction in https://github.com/getsentry/snuba/pull/1387
which splits `get_entity()` and `get_default_entity()`.

I tried to keep this set of changes as small (and readable) as possible,
there is further refactoring required here that can be completed as a
follow up once this approach has been reviewed.