fix(discover) Add a translator for functions on NULL (#1383)
When an impossible query is generated, the current translator will change any
non-existent columns to NULL. However, if that column is inside a function, it
can cause exceptions to happen if the function is called on a NULL value. If a
function is being called and all of its parameters are NULL, change the entire
function to be `ifNull(NULL, NULL)` so that any errors are avoided.