feat: Store entity name in logical query
This change involves storing the entity name in the logical query,
and using that entity from the query in order to run the query pipeline,
instead of relying on the dataset's default entity. Dataset no longer has a
`get_query_plan_builder` or `get_query_processors` method, this should
always be used from the entity instead.

The Dataset now has a `select_entity` method, which allows a dataset to
provide custom entity selection behavior, so a non default entity can be
picked. Eventually, this entity selection will be done by Sentry and specified
as part of the SnQL query.

Most of these changes are included in https://github.com/getsentry/snuba/pull/1407,
I'm moving the non Discover specific changes that relate to query processing here
to hopefully lkeep things more focussed and easier to review.