feat(discover): Move Discover storage selection to entity selection (#1407)
This moves the Discover storage selection decision up one level making
it the responsibility of the Discover dataset to perform entity selection.
This is closer to how Discover will eventually work - a specific entity
(errors, transactions, or both) will need to be specified by the user
when using SnQL. In order to use SnQL, this logic would also need to be
ported to Sentry.

This change involves storing the entity name in the logical query,
and using that entity from the query in order to run the query pipeline,
instead of relying on the dataset's default entity. Dataset no longer has a
get_query_plan_builder or get_query_processors method, this should
always be used from the entity instead.

The Dataset now has a select_entity method, which allows a dataset to
provide custom entity selection behavior, so a non default entity can be
picked. Eventually, this entity selection will be done by Sentry and specified
as part of the SnQL query.