feat: Align the Events and Discover column mapping
This aligns the columns of events and Discover so that the Discover
fields "release", "user" and "dist" are now supported on the Events
entity.

Ideally, these will be used by Sentry instead of `tags[sentry:release]`,
`tags[sentry:user]` and `tags[sentry:dist]` in the API, but both will
remain supported.

Note that this mapping won't be required when we switch to the errors
entity since "release", "user" and "dist" are already the same as the
underlying column names in errors.

The only remaining differences now between the Discover and Events
entity should relate to Null mapping of fields that are only present in
transactions.