fix(discover): Additional handling for date time literals
Conditions on the timestamp can be generated from discover as
`["less", ["timestamp", "2020-01-01T00:00:00.000000Z"]]` in a boolean search.
However, only top level conditions using infix operators have been special
cased to parse the date time literal as a Literal whereas this is parsed as a
Column when nested in a boolean search. This change updates the timeseries
processor in order to match and process this particular use of date time
literals correctly.