feat(query) Getting rid of aliases step 2a: Refactor web/query to prepare for using SelectedExpression objects (#1148)
Full context in the description of #1057 (see the "Overall plan")

This is a part of the step 2 according to that plan. Step 2 means, after the query runs, instead of relying on aliases in nAST expression to provide provide the column names the user expects in the output, we will take them from the SelectedExpression created in #1057.
This means we will be able to use internal aliases during the query processing avoiding any possible shadowing. And after that, we will be able to remove aliases alltogether.

This PR does some refactoring of web/query to create the place where we will process the query output, and ensures we start from a sane state by logging any discrepancy between SelectedClause names and existing aliases.